# openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogEntryOut": {
        "description": "The catalogue as clients see it.\n\nThe mapping lives here rather than on the catalogue entry: the wire\nshape is an HTTP concern, and `api/storage` is not allowed to know\none. A `describe()` on the entry would have made the storage layer\nthe place where a field rename in the contract has to be made.",
        "properties": {
          "descending": {
            "title": "Descending",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "family": {
            "title": "Family",
            "type": "string"
          },
          "filters": {
            "items": {
              "type": "string"
            },
            "title": "Filters",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "sort_key": {
            "items": {
              "type": "string"
            },
            "title": "Sort Key",
            "type": "array"
          },
          "symbol_scoped": {
            "title": "Symbol Scoped",
            "type": "boolean"
          },
          "table": {
            "title": "Table",
            "type": "string"
          }
        },
        "required": [
          "name",
          "family",
          "scope",
          "kind",
          "table",
          "sort_key",
          "descending",
          "filters",
          "symbol_scoped",
          "description"
        ],
        "title": "CatalogEntryOut",
        "type": "object"
      },
...
      "Collection_CatalogEntryOut_": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntryOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[CatalogEntryOut]",
        "type": "object"
      },
...
      "Collection_dict_str__Any__": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When this data was last verified against the source. Null where the schema keeps no fetch timestamp -- the price tables deliberately do not, since a per-row timestamp would cost gigabytes to answer a question the bar's own ts_utc already covers.",
            "title": "As Of"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` to fetch the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "Collection[dict[str, Any]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "clientCredentials": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "bars:read": "Read bars data",
              "discovery:read": "Read discovery data",
              "domains:read": "Read domains data",
              "fundamentals:read": "Read fundamentals data",
              "holders:read": "Read holders data",
              "news:read": "Read news data",
              "reference:read": "Read reference data"
            },
            "tokenUrl": "/oauth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Read-only access to the yfin market data warehouse. Authenticate with the OAuth2 client credentials flow.",
    "title": "yfin Data API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/datasets": {
      "get": {
        "description": "What this token can read, by default.\n\nFiltering to the caller's scopes is the default because advertising\ndata they cannot fetch is noise. `?all=true` is the documented way to\nsee the rest -- a deliberate choice, not an accident of implementation.",
        "operationId": "list_datasets_v1_datasets_get",
        "parameters": [
          {
            "description": "List every dataset, not only those this token can read.",
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "description": "List every dataset, not only those this token can read.",
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_CatalogEntryOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": []
          }
        ],
        "summary": "Dataset catalogue",
        "tags": [
          "datasets"
        ]
      }
    },
    "/v1/datasets/{name}": {
      "get": {
        "description": "Rows from one dataset.\n\nAny query parameter that is not reserved is treated as a filter, and\nonly the columns the dataset declared as filterable are accepted --\nan unknown one is refused rather than ignored, because silently\ndropping a filter returns more data than the caller asked for and\nlooks like it worked.",
        "operationId": "read_dataset_v1_datasets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "clientCredentials": []
          }
        ],
        "summary": "Dataset rows",
        "tags": [
          "datasets"
        ]
      }
    },
...
  }
}